devtrack_server/pyproject.toml
[project]
name = "ai-powered-task-automation"
version = "1.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "aiohttp>=3.12.13",
    "azure-identity>=1.23.0",
    "duckdb>=1.3.1",
    "fastapi[standard]>=0.115.12",
    "msgraph-sdk>=1.35.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "pygithub>=2.6.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    # NLP & AI Dependencies (Critical for Phase 3)
    "spacy>=3.7.0",
    "en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl",
    "sentence-transformers>=2.2.0",
    "ollama>=0.1.0",
    # Task Matching & Parsing
    "fuzzywuzzy>=0.18.0",
    "python-Levenshtein>=0.21.0",
    "dateparser>=1.2.0",
    # Testing Framework
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pandas-stubs==2.3.2.250926",
]

[project.optional-dependencies]
openai = ["openai>=1.0.0"]
anthropic = ["anthropic>=0.25.0"]
cloud = ["openai>=1.0.0", "anthropic>=0.25.0"]
mongodb = ["motor>=3.3.0"]
